P4-D: modularize today_feed_interactions into handler/builder/feedback widgets; remove size-exempt file and update docs
